+++ lib/src/modules/login/widgets/console_login_widget.dart
@@ -9,6 +9,7 @@ import 'package:flutter_trc/src/libraries/shared_prefrences/app_prefrences.dart'
 import 'package:flutter_trc/src/modules/login/resources/collector_user_controller.dart';
 import 'package:flutter_trc/src/modules/login/resources/login_types.dart';
 import 'package:flutter_trc/src/resources/user_details.dart';
+import 'package:flutter_trc/src/utils/app_util.dart';
 import 'package:provider/provider.dart';
 
 import '../l10n.dart';
@@ -30,12 +31,14 @@ class ConsoleLoginWidget extends StatelessWidget {
             serviceVersion: "v1",
             enterMobileNumberTitle: loginType == LoginTypes.rmsLogin ? "RMS Login" : "Shipex Login",
             companyName: "cashify",
+            otpEventKey: "otp_oms_event_active",
             isCompanyNameEditable: true,
             enterMobileNumberSubTitle: l10n.pleaseEnterCredentials,
-            mfaBypassClientId: "sales-rms:epoch",
+            mfaBypassClientId: getClientId(),
+            // mfaBypassClientId: "sales-rms:epoch",
             versionNumber: "",
             loginType: LoginType.mobile,
-            onSubmit: (userAuth, mobileNumber, mode, pin, {companyKey}) async {
+            onSubmit: (userAuth, mobileNumber, mode, pin, {required companyKey}) async {
               Logger.debug('mydebug-----_RmsLoginWidgetState.build', [
                 'userAuth',
                 userAuth,