+++ CHANGELOG.md
@@ -12,14 +12,14 @@
 
 # v3.4.2+56 (09-10-2024)
 
-### HotFix:
+### Hotfix:
 
-* `Correct integration of MFA bypass client ID`[(SCM-7761)](https://reglobe.atlassian.net/browse/SCM-7761)
+* `Change client key for console login`[(SCM-7761)](https://reglobe.atlassian.net/browse/SCM-7761)
 
 
 # v3.4.1+55 (08-10-2024)
 
-### HotFix:
+### Hotfix and Improvements:
 
 * `Email based otp from console`[(SCM-7716)](https://reglobe.atlassian.net/browse/SCM-7716)
 * `QC/TRC - Manual Testing - Brand is not auto-picked at the time of testing.`[(SCM-7711)](https://reglobe.atlassian.net/browse/SCM-7711)